M8 - FullStack/simple-flask-app/requirements.txt
bcrypt==3.2.0
cffi==1.14.6
click==8.0.1
Flask==2.0.1
greenlet==1.1.0
gunicorn==20.1.0
itsdangerous==2.0.1
Jinja2==3.0.1
MarkupSafe==2.0.1
psycopg2-binary==2.9.1
pycparser==2.20
PyJWT==2.1.0
python-dotenv==0.18.0
six==1.16.0
SQLAlchemy==1.4.20
Werkzeug==2.0.1
